+++ chat-widget/src/components/livechatwidget/common/defaultProps/dummyDefaultProps.ts
@@ -15,6 +15,7 @@ import { createAvatarMiddleware } from "../../../webchatcontainerstateful/webcha
 import { createMarkdown } from "../createMarkdown";
 import { groupActivitiesMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/groupActivitiesMiddleware";
 import { typingIndicatorMiddleware } from "../../../webchatcontainerstateful/webchatcontroller/middlewares/renderingmiddlewares/typingIndicatorMiddleware";
+import { createWebChatTelemetry } from "../../../webchatcontainerstateful/webchatcontroller/webchattelemetry/WebChatLogger";
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export const dummyDefaultProps: ILiveChatWidgetProps = {
@@ -1374,7 +1375,8 @@ export const dummyDefaultProps: ILiveChatWidgetProps = {
         },
         authClientFunction: undefined,
         isReconnectEnabled: undefined,
-        reconnectId: undefined
+        reconnectId: undefined,
+        redirectInSameWindow: undefined
     },
     styleProps: {
         generalStyles: {
@@ -1445,7 +1447,7 @@ export const dummyDefaultProps: ILiveChatWidgetProps = {
             internalErrorBoxClass: undefined,
             internalRenderErrorBox: undefined,
             locale: "en-US",
-            onTelemetry: undefined,
+            onTelemetry: createWebChatTelemetry(),
             overrideLocalizedStrings: undefined,
             renderMarkdown: createMarkdown(false, false),
             scrollToEndButtonMiddleware: undefined,