+++ chat-widget/src/components/livechatwidget/common/reconnectChatHelper.ts
@@ -1,10 +1,10 @@
 import "regenerator-runtime/runtime";
 
-import { LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
-
-import { Constants } from "../../../common/Constants";
+import { BroadcastEvent, LogLevel, TelemetryEvent } from "../../../common/telemetry/TelemetryConstants";
+import { BroadcastService } from "@microsoft/omnichannel-chat-components";
 import { ConversationState } from "../../../contexts/common/ConversationState";
 import { Dispatch } from "react";
+import { ICustomEvent } from "@microsoft/omnichannel-chat-components/lib/types/interfaces/ICustomEvent";
 import { ILiveChatWidgetAction } from "../../../contexts/common/ILiveChatWidgetAction";
 import { ILiveChatWidgetProps } from "../interfaces/ILiveChatWidgetProps";
 import { IReconnectChatContext } from "../../reconnectchatpanestateful/interfaces/IReconnectChatContext";
@@ -49,24 +49,85 @@ const getReconnectIdForAuthenticatedChat = async (props: ILiveChatWidgetProps, c
 };
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-const handleUnauthenticatedReconnectChat = async (dispatch: Dispatch<ILiveChatWidgetAction>, reconnectId: string, initStartChat: any) => {
-    const reconnectAvailabilityResponse: IReconnectChatContext = await getChatReconnectContext(reconnectId);
-    if (reconnectAvailabilityResponse && reconnectAvailabilityResponse.redirectURL) {
-        redirectPage(reconnectAvailabilityResponse.redirectURL);
+const handleUnauthenticatedReconnectChat = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, reconnectId: string, initStartChat: any, redirectInSameWindow: boolean | undefined) => {
+    const reconnectAvailabilityResponse: IReconnectChatContext = await getChatReconnectContext(chatSDK, reconnectId);
+    if (shouldRedirectOrStartNewChat(reconnectAvailabilityResponse)) {
+        await redirectOrStartNewChat(reconnectAvailabilityResponse, chatSDK, dispatch, setAdapter, initStartChat, redirectInSameWindow);
     } else {
-        const optionalParams = { reconnectId: reconnectId };
-        dispatch({ type: LiveChatWidgetActionType.SET_RECONNECT_ID, payload: reconnectId });
-        dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Loading });
-        await initStartChat(optionalParams);
+        await setReconnectIdAndStartChat(chatSDK, dispatch, setAdapter, reconnectId, initStartChat);
     }
 };
 
-const redirectPage = (newUrl: string) => {
-    const data = {
-        messageName: Constants.redirectPageRequest,
-        newUrl: newUrl
-    };
-    window.parent.postMessage(data, "*");
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const startUnauthenticatedReconnectChat = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, reconnectId: string, initStartChat: any) => {
+    const reconnectAvailabilityResponse: IReconnectChatContext = await getChatReconnectContext(chatSDK, reconnectId);
+    if (!shouldRedirectOrStartNewChat(reconnectAvailabilityResponse)) {
+        await setReconnectIdAndStartChat(chatSDK, dispatch, setAdapter, reconnectId, initStartChat);
+    }
 };
 
-export { getReconnectIdForAuthenticatedChat, handleUnauthenticatedReconnectChat };
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const setReconnectIdAndStartChat = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, reconnectId: string, initStartChat: any) => {
+    const startUnauthenticatedReconnectChat: ICustomEvent = {
+        eventName: BroadcastEvent.StartUnauthenticatedReconnectChat,
+    };
+    BroadcastService.postMessage(startUnauthenticatedReconnectChat);
+    const optionalParams = { reconnectId: reconnectId };
+    dispatch({ type: LiveChatWidgetActionType.SET_RECONNECT_ID, payload: reconnectId });
+    dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Loading });
+    await initStartChat(chatSDK, dispatch, setAdapter, optionalParams);
+};
+
+const redirectPage = (redirectURL: string, redirectInSameWindow: boolean | undefined) => {
+    const redirectPageRequest: ICustomEvent = {
+        eventName: BroadcastEvent.RedirectPageRequest,
+        payload: {
+            redirectURL: redirectURL
+        }
+    };
+    BroadcastService.postMessage(redirectPageRequest);
+    if (redirectInSameWindow) {
+        window.location.href = redirectURL;
+    }
+};
+
+const shouldRedirectOrStartNewChat = (reconnectAvailabilityResponse: IReconnectChatContext) => {
+    return reconnectAvailabilityResponse && !reconnectAvailabilityResponse.reconnectId;
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const startNewChatEmptyRedirectionUrl = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, initStartChat: any) => {
+    const startUnauthenticatedReconnectChat: ICustomEvent = {
+        eventName: BroadcastEvent.StartUnauthenticatedReconnectChat,
+    };
+    BroadcastService.postMessage(startUnauthenticatedReconnectChat);
+    // Getting PreChat Survey Context
+    const parseToJson = false;
+    const preChatSurveyResponse: string = await chatSDK.getPreChatSurvey(parseToJson);
+    if (preChatSurveyResponse) {
+        dispatch({ type: LiveChatWidgetActionType.SET_PRE_CHAT_SURVEY_RESPONSE, payload: preChatSurveyResponse });
+        dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Prechat });
+    } else {
+        dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Loading });
+        await initStartChat(chatSDK, dispatch, setAdapter);
+    }
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const handleRedirectUnauthenticatedReconnectChat = async (chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, initStartChat: any, reconnectId: string, redirectInSameWindow: boolean | undefined) => {
+    const reconnectAvailabilityResponse: IReconnectChatContext = await getChatReconnectContext(chatSDK, reconnectId);
+    if (shouldRedirectOrStartNewChat(reconnectAvailabilityResponse)) {
+        await redirectOrStartNewChat(reconnectAvailabilityResponse, chatSDK, dispatch, setAdapter, initStartChat, redirectInSameWindow);
+    }
+};
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const redirectOrStartNewChat = async (reconnectAvailabilityResponse: IReconnectChatContext, chatSDK: any, dispatch: Dispatch<ILiveChatWidgetAction>, setAdapter: any, initStartChat: any, redirectInSameWindow: boolean | undefined) => {
+    if (reconnectAvailabilityResponse.redirectURL) {
+        redirectPage(reconnectAvailabilityResponse.redirectURL, redirectInSameWindow);
+    } else {
+        await startNewChatEmptyRedirectionUrl(chatSDK, dispatch, setAdapter, initStartChat);
+    }
+};
+
+export { getChatReconnectContext, getReconnectIdForAuthenticatedChat, handleUnauthenticatedReconnectChat, startUnauthenticatedReconnectChat, handleRedirectUnauthenticatedReconnectChat };